fix(capture): the meeting hold is meetings-only, and the gates that hold it
Three corrections to the change above, each from a gate or a lane that
disagreed with it.

The discovery membership arm is restricted to meetings. Capture stamps
the acting seat as a participant of EVERY activity it writes, so the
unrestricted arm let a connector's granting human discover any row their
own connector had captured — including one later filed under a record
they may not read, which capture deliberately refuses to replay onto.

An agent-created person starts owner-scoped; a CONNECTOR-created one is
left to the path that creates it. People's own ensurer already decides
that for itself, and overruling it hid the person who booked a public
meeting from everyone but one seat.

The stranded-contacts selector asks for its evidence before the page
limit, so a meeting-only contact the verdict engine cannot judge no
longer consumes a slot and crowds out one it can.

The two stored-original passes now share one drain. Four files that
crossed a size cap are split by concept, and the registers that key on
file paths follow them.
