+++ backend/danglingcitations.txt
@@ -47,7 +47,6 @@ backend/gates/piicoverage_test.go core 0244
 backend/gates/piicoverage_test.go core 0245
 backend/gates/rbacgate_test.go core 0217
 backend/gates/rlsclaims_test.go core 0217
-backend/gates/tableownership_test.go migration 0219
 backend/gates/tableownershipwaivers_test.go core 0245
 backend/internal/compose/accountdraft/input.go core 0203
 backend/internal/compose/agentgate_pin_test.go core 0063
