+++ backend/gates/audiencereaders_test.go
@@ -100,6 +100,7 @@ var audienceLiteralMarkers = []*regexp.Regexp{
 // covers one reader and not every reader added to that file afterwards.
 var audienceReadersAdmitted = gatekit.Waive(map[string]string{
 	"internal/compose/participantreplay.go:selectReplayCandidates":                   "the replay pass re-parses a stored provider original to fill in the participant rows an earlier capture missed, as the system principal, and every column it reads dies inside the transaction: the payload is handed to the mail parser and what SURVIVES is activity_participant rows — addresses and roles, never text. It cannot compose the audience because it is the pass that establishes who the participants ARE, which is one of the inputs the audience arm reads; gating it on the answer it computes would leave a message whose participants were missed permanently unreadable by the very people who were on it. The cost is that a limited message's stored original is parsed in-process by a system pass",
+	"internal/compose/meetingattendeerepair.go:selectMeetingRepairCandidates":        "the third sibling of the replay above, asking the same question of the same stored calendar originals under a corrected rule: which attendees are COLLEAGUES, now that a provider's attendee list is trusted to bind a user_id. It runs as the system principal over calendar resources only — never mail — and what survives the transaction is activity_participant rows: addresses, roles and a display name, never subject or body. It cannot compose the audience for the replay's reason in its sharpest form: this pass establishes the participant rows the audience arm READS, so gating it on that answer would leave a held meeting permanently unreadable by the very colleagues who were invited to it — which is the whole defect the pass exists to repair. The cost is that a held meeting's stored invitation is parsed in-process by a system pass",
 	"internal/compose/participantnamerecover.go:selectNameRecoveryCandidates":        "the sibling of the replay above, asking the other question of the same stored originals: not who was in this meeting but what the invitation CALLED them. It runs as the system principal, reads a calendar event resource — never mail — and what survives the transaction is a display name on a participant row and, where the guards admit it, a first and last name on a person nobody had named. No subject, no body, no text of any kind reaches a caller. It cannot compose the audience for the replay's reason and one narrower: an attendee's own name would be withheld from the very people on the meeting, and a contact called by the local part of their email would stay that way permanently. The cost is that a limited meeting's stored invitation is parsed in-process by a system pass",
 	"internal/modules/activities/capturenoise.go:Store.RedactCapturedNoiseTx":        "the noise redaction reads subject, body and raw to answer ONE question — is there content left to destroy — and its only effect is to null them. A gate here would exempt limited mail from destruction, which is the opposite of what the audience protects: it would leave a colleague's held message stored forever because nobody may read it. The cost is nil; no column it reads reaches a caller",
 	"internal/modules/privacy/retentionselectors.go:AuthorableScopes":                "the retention selectors name which activities a policy DESTROYS, and the content columns in them are tests for whether there is content left to destroy (`a.body IS NOT NULL`), never a projection: the query returns ids. Gating them on the audience would exempt limited mail from retention — a colleague's held message kept forever because nobody may read it, which is the opposite of what the audience protects. Same obligation as the noise redaction above, at the other place content is destroyed",
