+++ backend/internal/compose/attention/reach.go
@@ -170,7 +170,7 @@ func categoryOfSource(source crmcontracts.WorklistItemSource) crmcontracts.Workl
 		return categoryLeads
 	case sourceAtRisk, "brief_item":
 		return "deals_at_risk"
-	case "meeting", "meeting_outcome":
+	case sourceMeeting, sourceMeetingOutcome:
 		return "meetings"
 	case sourceTask, "conversation_claim":
 		return "tasks"
